The game will help your kids to learn and grow.
This game contains 598 different animals name, which I will keep on increasing. This is a single player version where you will be learn telling the name of the animals (all kind). and the next alphabet given to you is the last letter of your previous answer, which is making it more fun to play.